### 2020-10-15 #63 @whedon
[PRE REVIEW]: GeometricFlux.jl: a geometric deep learning library in Julia
**Submitting author:** @yuehhua (<a href="http://orcid.org/0000-0002-1554-1753">Yueh-Hua Tu</a>)
**Repository:** <a href="https://github.com/yuehhua/GeometricFlux.jl" target ="_blank">https://github.com/yuehhua/GeometricFlux.jl</a>
**Version:**
