### 2020-10-23 5115e9b
Avoid converting in GCNConv
