### 2020-10-23 b75bda5
Convert only when CuArray for all conv layers
