+++ src/GeometricFlux.jl
@@ -3,12 +3,13 @@ module GeometricFlux
 using Statistics: mean
 using SparseArrays: SparseMatrixCSC
 using LinearAlgebra: Adjoint, norm
+using Reexport
 
 using FillArrays: Fill
 using Flux
 using Flux: glorot_uniform, leakyrelu, GRUCell
 using Flux: @functor
-using GraphSignals
+@reexport using GraphSignals
 using LightGraphs
 using Requires
 using ScatterNNlib
