+++ src/GeometricFlux.jl
@@ -77,7 +77,8 @@ export
     generate_cluster,
 
     # utils
-    topk_index
+    topk_index,
+    bypass_graph
 
 const IntOrTuple = Union{Integer,Tuple}
 
