### 2020-11-02 37b76ee
Adjust edge_index_table API for directed
Fix
