+++ src/graph/index.jl
@@ -53,20 +53,31 @@ end
 Zygote.@nograd vertex_pair_table
 
 """
-    edge_index_table(adj[, num_E])
+    edge_index_table(adj[, directed])
 
 Generate a mapping from vertex pair (i, j) to edge index. The edge indecies are determined by
 the sorted vertex indecies.
 """
-function edge_index_table(adj::AbstractVector{<:AbstractVector{<:Integer}},
-                          num_E=sum(map(length, adj)))
+function edge_index_table(adj::AbstractVector{<:AbstractVector{<:Integer}}, directed::Bool=is_directed(adj))
     table = Dict{Tuple{UInt32,UInt32},UInt64}()
     e = one(UInt64)
-    for (i, js) = enumerate(adj)
-        js = sort(js)
-        for j = js
-            table[(i, j)] = e
-            e += one(UInt64)
+    if directed
+        for (i, js) = enumerate(adj)
+            js = sort(js)
+            for j = js
+                table[(i, j)] = e
+                e += one(UInt64)
+            end
+        end
+    else
+        for (i, js) = enumerate(adj)
+            js = sort(js)
+            js = js[i .≤ js]
+            for j = js
+                table[(i, j)] = e
+                table[(j, i)] = e
+                e += one(UInt64)
+            end
         end
     end
     table
@@ -79,6 +90,8 @@ function edge_index_table(vpair::AbstractVector{<:Tuple})
     end
     table
 end
+
+edge_index_table(fg::FeaturedGraph) = edge_index_table(fg.graph, fg.directed)
 
 Zygote.@nograd edge_index_table
 
