### 2020-11-02 ed1a127
Improve differentiability test
