+++ test/layers/conv.jl
@@ -21,8 +21,12 @@ adj = [0. 1. 0. 1.;
             Y = gc(X)
             @test size(Y) == (out_channel, N)
             
-            g = Zygote.gradient(() -> sum(gc(X)), params(gc))
-            @test length(g.grads) == 2
+            g = Zygote.gradient(x -> sum(gc(x)), X)[1]
+            @test size(g) == size(X)
+
+            g = Zygote.gradient(model -> sum(model(X)), gc)[1]
+            @test size(g.weight) == size(gc.weight)
+            @test size(g.bias) == size(gc.bias)
         end
 
         @testset "layer without graph" begin
@@ -36,8 +40,12 @@ adj = [0. 1. 0. 1.;
             @test size(node_feature(fg_)) == (out_channel, N)
             @test_throws AssertionError gc(X)
 
-            g = Zygote.gradient(() -> sum(node_feature(gc(fg))), params(gc))
-            @test length(g.grads) == 4
+            g = Zygote.gradient(x -> sum(node_feature(gc(x))), fg)[1]
+            @test size(g[].nf) == size(X)
+
+            g = Zygote.gradient(model -> sum(node_feature(model(fg))), gc)[1]
+            @test size(g.weight) == size(gc.weight)
+            @test size(g.bias) == size(gc.bias)
         end
     end
 
@@ -57,8 +65,12 @@ adj = [0. 1. 0. 1.;
             Y = cc(X)
             @test size(Y) == (out_channel, N)
 
-            g = Zygote.gradient(() -> sum(cc(X)), params(cc))
-            @test length(g.grads) == 3
+            g = Zygote.gradient(x -> sum(cc(x)), X)[1]
+            @test size(g) == size(X)
+
+            g = Zygote.gradient(model -> sum(model(X)), cc)[1]
+            @test size(g.weight) == size(cc.weight)
+            @test size(g.bias) == size(cc.bias)
         end
 
         @testset "layer without graph" begin
@@ -75,8 +87,12 @@ adj = [0. 1. 0. 1.;
             @test size(node_feature(fg_)) == (out_channel, N)
             @test_throws AssertionError cc(X)
 
-            g = Zygote.gradient(() -> sum(node_feature(cc(fg))), params(cc))
-            @test length(g.grads) == 4
+            g = Zygote.gradient(x -> sum(node_feature(cc(x))), fg)[1]
+            @test size(g[].nf) == size(X)
+
+            g = Zygote.gradient(model -> sum(node_feature(model(fg))), cc)[1]
+            @test size(g.weight) == size(cc.weight)
+            @test size(g.bias) == size(cc.bias)
         end
     end
 
@@ -92,9 +108,13 @@ adj = [0. 1. 0. 1.;
             Y = gc(X)
             @test size(Y) == (out_channel, N)
 
-            # Test that the gradient can be computed
-            g = Zygote.gradient(() -> sum(gc(X)), params(gc))
-            @test length(g.grads) == 5
+            g = Zygote.gradient(x -> sum(gc(x)), X)[1]
+            @test size(g) == size(X)
+
+            g = Zygote.gradient(model -> sum(model(X)), gc)[1]
+            @test size(g.weight1) == size(gc.weight1)
+            @test size(g.weight2) == size(gc.weight2)
+            @test size(g.bias) == size(gc.bias)
         end
 
         @testset "layer without graph" begin
@@ -110,8 +130,13 @@ adj = [0. 1. 0. 1.;
             @test size(node_feature(fg_)) == (out_channel, N)
             @test_throws AssertionError gc(X)
 
-            g = Zygote.gradient(() -> sum(node_feature(gc(fg))), params(gc))
-            @test length(g.grads) == 5
+            g = Zygote.gradient(x -> sum(node_feature(gc(x))), fg)[1]
[…]
