### 2020-11-02 f11fd11
Improve differentiability test for cuda
