### 2020-11-03 a4d7946
Refactor propagate into a generic version
