### 2020-11-03 bd3a3b4
Bug with differentiating GRUCell with CuArray
