### 2020-11-03 bbfd462
Refactor message passing
