### 2020-11-03 ddb14ea
Refactor graph net
