### 2020-11-03 bc2ec8d
Refactor syntax
