+++ src/layers/gn.jl
@@ -10,18 +10,15 @@ abstract type GraphNet end
 @inline function update_batch_edge(gn::T, adj, E, V, u) where {T<:GraphNet}
     n = size(adj, 1)
     edge_idx = edge_index_table(adj)
-    E_ = [apply_batch_message(gn, i, adj[i], edge_idx, E, V, u) for i in 1:n]
-    hcat(E_...)
+    hcat([apply_batch_message(gn, i, adj[i], edge_idx, E, V, u) for i in 1:n]...)
 end
 
 @inline function apply_batch_message(gn::T, i, js, edge_idx, E, V, u) where {T<:GraphNet}
-    E_ = [update_edge(gn, get_feature(E, edge_idx[(i,j)]), get_feature(V, i), get_feature(V, j), u) for j = js]
-    hcat(E_...)
+    hcat([update_edge(gn, get_feature(E, edge_idx[(i,j)]), get_feature(V, i), get_feature(V, j), u) for j = js]...)
 end
 
 @inline function update_batch_vertex(gn::T, Ē, V, u) where {T<:GraphNet}
-    V_ = [update_vertex(gn, get_feature(Ē, i), get_feature(V, i), u) for i = 1:size(V,2)]
-    hcat(V_...)
+    hcat([update_vertex(gn, get_feature(Ē, i), get_feature(V, i), u) for i = 1:size(V,2)]...)
 end
 
 @inline function aggregate_neighbors(gn::T, aggr::Symbol, E, accu_edge) where {T<:GraphNet}
@@ -63,14 +60,11 @@ end
 
 function propagate(gn::T, adj::AbstractVector{S}, E::R, V::Q, u::P,
                    naggr=nothing, eaggr=nothing, vaggr=nothing) where {T<:GraphNet,S<:AbstractVector,R,Q,P}
-    accu_edge = accumulated_edges(adj)
-
     E = update_batch_edge(gn, adj, E, V, u)
-    Ē = aggregate_neighbors(gn, naggr, E, accu_edge)
+    Ē = aggregate_neighbors(gn, naggr, E, accumulated_edges(adj))
     V = update_batch_vertex(gn, Ē, V, u)
     ē = aggregate_edges(gn, eaggr, E)
     v̄ = aggregate_vertices(gn, vaggr, V)
     u = update_global(gn, ē, v̄, u)
-
     E, V, u
 end
