### 2020-11-03 b091973
Integration of propagate
