### 2020-11-06 cbb9a3c
Correct GATConv computation of concat=false
