+++ src/layers/conv.jl
@@ -301,8 +301,12 @@ end
 
 # The same as update function in batch manner
 function update_batch_vertex(g::GATConv, M::AbstractMatrix, X::AbstractMatrix, u)
-    g.concat || (M = mean(M, dims=2))
-    return M .+ g.bias
+    M = M .+ g.bias
+    if !g.concat
+        N = size(M, 2)
+        M = reshape(mean(reshape(M, :, g.heads, N), dims=2), :, N)
+    end
+    return M
 end
 
 function (gat::GATConv)(X::AbstractMatrix)
