+++ src/cuda/conv.jl
@@ -5,6 +5,9 @@ function update_batch_vertex(g::GATConv, M::AbstractMatrix, X::CuMatrix, u)
 end
 
 function update_batch_vertex(g::GATConv, M::CuMatrix, X::CuMatrix, u)
-    g.concat || (M = mean(M, dims=2))
-    return M .+ g.bias
+    M = M .+ g.bias
+    if !g.concat
+        N = size(M, 2)
+        M = reshape(mean(reshape(M, :, g.heads, N), dims=2), :, N)
+    end
 end
