### 2020-11-09 2b095c3
Fix GCN on CPU
