### 2020-11-09 ab3a366
Remove duplicate softmax
