### 2020-11-10 5bd430c
Update name of scaled_laplacian matrix
