+++ src/layers/conv.jl
@@ -50,13 +50,12 @@ end
 
 function (g::GCNConv)(A::AbstractMatrix, X::AbstractMatrix)
     L̃ = normalized_laplacian(A, eltype(X); selfloop=true)
-
-    if isa(X, CuArray)
-        L̃ = convert(typeof(X), L̃)  # ensure L has the same type as X, especially X::CuArray
-    end
+    L̃ = convert(typeof(X), L̃)  # ensure L has the same type as X, especially X::CuArray
     
     g.σ.(g.weight * X * L̃ .+ g.bias)
 end
+
+(g::GCNConv)(A::AbstractMatrix, X::Transpose{S,R}) where {S,R<:AbstractMatrix} = g(A, R(X))
 
 function (g::GCNConv)(X::AbstractMatrix{T}) where {T}
     @assert has_graph(g.fg) "A GCNConv created without a graph must be given a FeaturedGraph as an input."
@@ -143,13 +142,12 @@ function (c::ChebConv)(X::AbstractMatrix{T}) where {T<:Real}
     @assert has_graph(c.fg) "A ChebConv created without a graph must be given a FeaturedGraph as an input."
     g = graph(c.fg)
     L̃ = scaled_laplacian(g, T)
-
-    if isa(X, CuArray)
-        L̃ = convert(typeof(X), L̃)  # ensure L̃ has the same type as X, especially X::CuArray
-    end
+    L̃ = convert(typeof(X), L̃)  # ensure L̃ has the same type as X, especially X::CuArray
     
     c(L̃, X)
 end
+
+(c::ChebConv)(X::Transpose{T,R}) where {T<:Real,R<:AbstractMatrix} = c(R(X))
 
 function (c::ChebConv)(fg::FeaturedGraph)
     @assert has_graph(fg) "A given FeaturedGraph must contain a graph."
@@ -159,10 +157,7 @@ function (c::ChebConv)(fg::FeaturedGraph)
     end
     X = node_feature(fg)
     L̃ = scaled_laplacian(adjacency_matrix(fg))
-
-    if isa(X, CuArray)
-        L̃ = convert(typeof(X), L̃)  # ensure L has the same type as X, especially X::CuArray
-    end
+    L̃ = convert(typeof(X), L̃)  # ensure L has the same type as X, especially X::CuArray
     
     X_ = c(L̃, X)
     FeaturedGraph(g, X_)
