### 2020-11-12 021837f
Fix softmaxing of alphas in GATConv
