### 2020-11-12 2677782
Testing GATConv with single vertices (without any neighbors)
