### 2020-11-12 70b7bdb
GATConv: add own vertex to cluster
