### 2020-11-14 25344ca
CompatHelper: bump compat for "CUDA" to "2.2"
