### 2020-11-14 383d6d7
Refactor inner product and shape of a in GATConv
