### 2020-11-15 938b639
Rename variable as paper used
