+++ src/layers/conv.jl
@@ -309,11 +309,11 @@ end
 # After some reshaping due to the multihead, we get the α from each message, 
 # then get the softmax over every α, and eventually multiply the message by α
 function apply_batch_message(g::GATConv, i, js, edge_idx, E::AbstractMatrix, X::AbstractMatrix, u)
-    alpha_messages = hcat([message(g, get_feature(X, i), get_feature(X, j), get_feature(E, edge_idx[(i,j)])) for j = js]...)
-    alpha_messages = reshape(alpha_messages, :, size(alpha_messages, 2)*g.heads)
-    alphas = Flux.softmax(alpha_messages[1, :])
+    e_ij = hcat([message(g, get_feature(X, i), get_feature(X, j), get_feature(E, edge_idx[(i,j)])) for j = js]...)
+    e_ij = reshape(e_ij, :, size(e_ij, 2)*g.heads)
+    alphas = Flux.softmax(e_ij[1, :])
     alphas = reshape(alphas, 1, :) # We need a line so that every α will multiply a column
-    messages = alpha_messages[2:end, :] .* alphas
+    messages = e_ij[2:end, :] .* alphas
     reshape(messages, size(messages, 1)*g.heads, :)
 end
 
