### 2020-11-15 e585dbe
Softmax by each head
