### 2020-11-15 2e360b4
Refactor apply_batch_message
