### 2020-11-15 dc5026c
Add add_self_loop!
