### 2020-11-18 9a5518a
match matrix type to prevent CUDA.CUSPARSE error
fixes "ERROR: LoadError: MethodError: no method matching CUDA.CUSPARSE.CuSparseMatrixCSC{Float32}(::CuArray{Float32,2})"
