### 2020-11-19 b2ec3c0
Allow Travis test on nightly
