+++ Project.toml
@@ -40,7 +40,7 @@ SimpleWeightedGraphs = "1.1"
 StaticArrays = "0.12"
 Zygote = "0.4 - 0.5"
 ZygoteRules = "0.2"
-julia = "1.4 - 1.5"
+julia = "1.4 - 1.6"
 
 [extras]
 Test = "8dfed614-e22c-5e08-85e1-65c5234f0b40"
