### 2020-11-28 4f6c342
CompatHelper: bump compat for "StaticArrays" to "1.0"
