+++ Project.toml
@@ -37,7 +37,7 @@ Reexport = "0.2"
 Requires = "1.0.0"
 ScatterNNlib = "0.1"
 SimpleWeightedGraphs = "1.1"
-StaticArrays = "0.12"
+StaticArrays = "0.12, 1.0"
 Zygote = "0.4 - 0.5"
 ZygoteRules = "0.2"
 julia = "1.4 - 1.6"
