### 2020-12-01 8ec4416
Remove deps StaticArrays
