+++ Project.toml
@@ -19,7 +19,6 @@ Requires = "ae029012-a4dd-5104-9daa-d747884805df"
 ScatterNNlib = "b1168b60-8710-48c1-88d2-5c53ae207dd0"
 SimpleWeightedGraphs = "47aef6b3-ad0c-573a-a1e2-d07658019622"
 SparseArrays = "2f01184e-e22b-5df5-ae63-d93ebab69eaf"
-StaticArrays = "90137ffa-7385-5640-81b9-e52037218182"
 Statistics = "10745b16-79ce-11e8-11f9-7d13ad32a3b2"
 Zygote = "e88e6eb3-aa80-5325-afca-941959d7151f"
 ZygoteRules = "700de1a5-db45-46bc-99cf-38207098b444"
@@ -37,7 +36,6 @@ Reexport = "0.2"
 Requires = "1.0.0"
 ScatterNNlib = "0.1"
 SimpleWeightedGraphs = "1.1"
-StaticArrays = "0.12"
 Zygote = "0.4 - 0.5"
 ZygoteRules = "0.2"
 julia = "1.4 - 1.6"
