### 2020-12-03 2583daa
Fix test failure caused by FeaturedGraph API change
