### 2020-12-04 6ce1b35
Cast testing data to Float32
