+++ test/layers/conv.jl
@@ -16,7 +16,8 @@ adj_single_vertex =   T[0. 0. 0. 1.;
 
 @testset "layer" begin
     @testset "GCNConv" begin
-        X = rand(in_channel, N)
+        X = rand(Float32, in_channel, N)
+        Xt = transpose(rand(Float32, N, in_channel))
         @testset "layer with graph" begin
             gc = GCNConv(adj, in_channel=>out_channel)
             @test size(gc.weight) == (out_channel, in_channel)
@@ -27,8 +28,7 @@ adj_single_vertex =   T[0. 0. 0. 1.;
             @test size(Y) == (out_channel, N)
             
             # Test with transposed features
-            Xt = rand(N, in_channel)
-            Y = gc(transpose(Xt))
+            Y = gc(Xt)
             @test size(Y) == (out_channel, N)
             
             g = Zygote.gradient(x -> sum(gc(x)), X)[1]
@@ -51,8 +51,7 @@ adj_single_vertex =   T[0. 0. 0. 1.;
             @test_throws AssertionError gc(X)
 
             # Test with transposed features
-            Xt = rand(N, in_channel)
-            fgt = FeaturedGraph(adj, nf=transpose(Xt))
+            fgt = FeaturedGraph(adj, nf=Xt)
             fgt_ = gc(fgt)
             @test size(node_feature(fgt_)) == (out_channel, N)
             
@@ -68,7 +67,8 @@ adj_single_vertex =   T[0. 0. 0. 1.;
 
     @testset "ChebConv" begin
         k = 6
-        X = rand(in_channel, N)
+        X = rand(Float32, in_channel, N)
+        Xt = transpose(rand(Float32, N, in_channel))
         @testset "layer with graph" begin
             cc = ChebConv(adj, in_channel=>out_channel, k)
             @test size(cc.weight) == (out_channel, in_channel, k)
@@ -82,8 +82,7 @@ adj_single_vertex =   T[0. 0. 0. 1.;
             @test size(Y) == (out_channel, N)
 
             # Test with transposed features
-            Xt = rand(N, in_channel)
-            Y = cc(transpose(Xt))
+            Y = cc(Xt)
             @test size(Y) == (out_channel, N)
             
             g = Zygote.gradient(x -> sum(cc(x)), X)[1]
@@ -109,8 +108,7 @@ adj_single_vertex =   T[0. 0. 0. 1.;
             @test_throws AssertionError cc(X)
 
             # Test with transposed features
-            Xt = rand(N, in_channel)
-            fgt = FeaturedGraph(adj, nf=transpose(Xt))
+            fgt = FeaturedGraph(adj, nf=Xt)
             fgt_ = cc(fgt)
             @test size(node_feature(fgt_)) == (out_channel, N)
             
@@ -124,6 +122,8 @@ adj_single_vertex =   T[0. 0. 0. 1.;
     end
 
     @testset "GraphConv" begin
+        X = rand(Float32, in_channel, N)
+        Xt = transpose(rand(Float32, N, in_channel))
         @testset "layer with graph" begin
             gc = GraphConv(adj, in_channel=>out_channel)
             @test graph(gc.fg) == [[2,4], [1,3], [2,4], [1,3]]
@@ -131,15 +131,11 @@ adj_single_vertex =   T[0. 0. 0. 1.;
             @test size(gc.weight2) == (out_channel, in_channel)
             @test size(gc.bias) == (out_channel,)
 
-            X = rand(in_channel, N)
             Y = gc(X)
             @test size(Y) == (out_channel, N)
 
-            
-
             # Test with transposed features
-            Xt = rand(N, in_channel)
-            Y = gc(transpose(Xt))
+            Y = gc(Xt)
             @test size(Y) == (out_channel, N)
             
[…]
