### 2020-12-06 ca8fcf8
Add Julia v1.4 to ci
