+++ .github/workflows/ci.yml
@@ -10,6 +10,7 @@ jobs:
       fail-fast: false
       matrix:
         version:
+          - '1.4'
           - '1.5'
           - 'nightly'
         os:
