+++ .github/workflows/ci.yml
@@ -62,8 +62,8 @@ jobs:
       - run: |
           julia --project=docs -e '
             using Documenter: doctest
-            using GraphSignals
-            doctest(GraphSignals)'
+            using GeometricFlux
+            doctest(GeometricFlux)'
       - run: julia --project=docs docs/make.jl
         env:
           GITHUB_TOKEN: ${{ secrets.GITHUB_TOKEN }}
