+++ Project.toml
@@ -1,7 +1,7 @@
 name = "GeometricFlux"
 uuid = "7e08b658-56d3-11e9-2997-919d5b31e4ea"
 authors = ["Yueh-Hua Tu <a504082002@gmail.com>"]
-version = "0.7.4"
+version = "0.7.5"
 
 [deps]
 CUDA = "052768ef-5323-5732-b1bb-66c8b64840ba"
@@ -24,7 +24,7 @@ Zygote = "e88e6eb3-aa80-5325-afca-941959d7151f"
 ZygoteRules = "700de1a5-db45-46bc-99cf-38207098b444"
 
 [compat]
-CUDA = "1.2 - 2.1, 2.2"
+CUDA = "1.2 - 2.2"
 DataStructures = "0.17 - 0.18"
 FillArrays = "0.8 - 0.10"
 Flux = "0.10 - 0.11"
