+++ Project.toml
@@ -36,7 +36,7 @@ Reexport = "0.2"
 Requires = "1.0.0"
 ScatterNNlib = "0.1"
 SimpleWeightedGraphs = "1.1"
-Zygote = "0.4 - 0.5"
+Zygote = "0.4 - 0.5, 0.6"
 ZygoteRules = "0.2"
 julia = "1.4 - 1.6"
 
