+++ Project.toml
@@ -32,7 +32,7 @@ GraphMLDatasets = "0.1"
 GraphSignals = "0.1"
 LightGraphs = "1.3"
 MetaGraphs = "0.6"
-Reexport = "0.2"
+Reexport = "0.2, 1.0"
 Requires = "1.0.0"
 ScatterNNlib = "0.1"
 SimpleWeightedGraphs = "1.1"
