### 2021-01-09 1d2439d
CompatHelper: bump compat for "CUDA" to "2.4"
