### 2021-01-14 99f4ded
CompatHelper: bump compat for "CUDA" to "2.5"
