### 2021-01-26 e8cbfc5
CompatHelper: bump compat for "CUDA" to "2.6"
