### 2021-02-09 #70 @ranjanan
@dhairyalgandhi look over JuliaCon guide and start your review.

Just as a first point, @yuehhua, I would improve the quality of plots in the paper. They do not look publication quality.
