### 2021-02-11 21fc4f3
Update figures
