+++ paper/paper.tex
@@ -58,19 +58,19 @@ JuliaGraph ecosystem.
 A github repository is available in https://github.com/yuehhua/GeometricFlux.jl.
 
 \begin{figure*}[t]
-\centerline{\includegraphics[width=18cm]{figures/scatter_add.png}}
+\centerline{\includegraphics[width=18cm]{figures/scatter_add.svg}}
 \caption{Benchmark for scatter add on CPU and GPU.}
     \label{fig:scatter_add}
 \end{figure*}
 
 \begin{figure*}[h]
-\centerline{\includegraphics[width=18cm]{figures/scatter_mean.png}}
+\centerline{\includegraphics[width=18cm]{figures/scatter_mean.svg}}
 \caption{Benchmark for scatter mean on CPU and GPU.}
     \label{fig:scatter_mean}
 \end{figure*}
 
 \begin{figure*}[h]
-\centerline{\includegraphics[width=18cm]{figures/scatter_max.png}}
+\centerline{\includegraphics[width=18cm]{figures/scatter_max.svg}}
 \caption{Benchmark for scatter max on CPU and GPU.}
     \label{fig:scatter_max}
 \end{figure*}
