+++ paper/paper.tex
@@ -3,6 +3,7 @@
 \documentclass{juliacon}
 \setcounter{page}{1}
 \usepackage{verbatim}
+\usepackage{svg}
 \usepackage{amsmath}
 
 \begin{document}
@@ -58,21 +59,27 @@ JuliaGraph ecosystem.
 A github repository is available in https://github.com/yuehhua/GeometricFlux.jl.
 
 \begin{figure*}[t]
-\centerline{\includegraphics[width=18cm]{figures/scatter_add.svg}}
-\caption{Benchmark for scatter add on CPU and GPU.}
+\centering{
+    \resizebox{180mm}{!}{\input{figures/scatter_add.pdf_tex}}
+    \caption{Benchmark for scatter add on CPU and GPU.}
     \label{fig:scatter_add}
+}
 \end{figure*}
 
 \begin{figure*}[h]
-\centerline{\includegraphics[width=18cm]{figures/scatter_mean.svg}}
-\caption{Benchmark for scatter mean on CPU and GPU.}
+\centering{
+    \resizebox{180mm}{!}{\input{figures/scatter_mean.pdf_tex}}
+    \caption{Benchmark for scatter mean on CPU and GPU.}
     \label{fig:scatter_mean}
+}
 \end{figure*}
 
 \begin{figure*}[h]
-\centerline{\includegraphics[width=18cm]{figures/scatter_max.svg}}
-\caption{Benchmark for scatter max on CPU and GPU.}
+\centering{
+    \resizebox{180mm}{!}{\input{figures/scatter_max.pdf_tex}}
+    \caption{Benchmark for scatter max on CPU and GPU.}
     \label{fig:scatter_max}
+}
 \end{figure*}
 
 \section{Extending Framework}
