+++ paper/paper.tex
@@ -59,27 +59,21 @@ JuliaGraph ecosystem.
 A github repository is available in https://github.com/yuehhua/GeometricFlux.jl.
 
 \begin{figure*}[t]
-\centering{
-    \resizebox{180mm}{!}{\input{figures/scatter_add.pdf_tex}}
-    \caption{Benchmark for scatter add on CPU and GPU.}
-    \label{fig:scatter_add}
-}
+\resizebox{180mm}{!}{\input{figures/scatter_add.pdf_tex}}
+\caption{Benchmark for scatter add on CPU and GPU.}
+\label{fig:scatter_add}
 \end{figure*}
 
-\begin{figure*}[h]
-\centering{
-    \resizebox{180mm}{!}{\input{figures/scatter_mean.pdf_tex}}
-    \caption{Benchmark for scatter mean on CPU and GPU.}
-    \label{fig:scatter_mean}
-}
+\begin{figure*}[ht]
+\resizebox{180mm}{!}{\input{figures/scatter_mean.pdf_tex}}
+\caption{Benchmark for scatter mean on CPU and GPU.}
+\label{fig:scatter_mean}
 \end{figure*}
 
-\begin{figure*}[h]
-\centering{
-    \resizebox{180mm}{!}{\input{figures/scatter_max.pdf_tex}}
-    \caption{Benchmark for scatter max on CPU and GPU.}
-    \label{fig:scatter_max}
-}
+\begin{figure*}[ht]
+\resizebox{180mm}{!}{\input{figures/scatter_max.pdf_tex}}
+\caption{Benchmark for scatter max on CPU and GPU.}
+\label{fig:scatter_max}
 \end{figure*}
 
 \section{Extending Framework}
