+++ Project.toml
@@ -27,7 +27,7 @@ ZygoteRules = "700de1a5-db45-46bc-99cf-38207098b444"
 CUDA = "1.2 - 2.6"
 DataStructures = "0.17 - 0.18"
 FillArrays = "0.8 - 0.11"
-Flux = "0.10 - 0.11"
+Flux = "0.10 - 0.11, 0.12"
 GraphMLDatasets = "0.1"
 GraphSignals = "0.1"
 LightGraphs = "1.3"
