+++ .gitlab-ci.yml
@@ -1,110 +1,0 @@
-variables:
-  CI_APT_INSTALL: 'libgomp1'
-  JULIA_NUM_THREADS: '4'
-  NVIDIA_VISIBLE_DEVICES: 'all'
-  NVIDIA_DRIVER_CAPABILITIES: 'compute,utility'
-  CI_THOROUGH: 'true'
-  JULIA_CUDA_VERSION: '10.2'
-  JULIA_CUDA_USE_BINARYBUILDER: 'true'
-  DATADEPS_ALWAYS_ACCEPT: 'true'
-
-include:
-  - 'https://raw.githubusercontent.com/JuliaGPU/gitlab-ci/master/templates/v6.yml'
-
-image: ubuntu:bionic
-
-test:v1.4:
-  extends:
-    - .julia:1.4
-    - .test
-  variables:
-    CI_VERSION_TAG: 'v1.4'
-  script:
-    - |
-      julia --project -e '
-        using Pkg; Pkg.add("PyCall");
-        Pkg.instantiate();
-        ENV["PYTHON"] = ""; Pkg.build();
-        using PyCall; pyimport_conda("scipy", "scipy")'
-    - julia --project -e 'using PyCall; pyimport_conda("scipy", "scipy")' # retry
-    - |
-      julia --project -e '
-        using Pkg;
-        Pkg.instantiate();
-        Pkg.build();
-        Pkg.test(; coverage=true);'
-  tags:
-    - nvidia
-    - cuda_10.2
-
-test:v1.5:
-  extends:
-    - .julia:1.5
-    - .test
-  variables:
-    CI_VERSION_TAG: 'v1.5'
-  script:
-    - |
-      julia --project -e '
-        using Pkg; Pkg.add("PyCall");
-        Pkg.instantiate();
-        ENV["PYTHON"] = ""; Pkg.build();
-        using PyCall; pyimport_conda("scipy", "scipy")'
-    - julia --project -e 'using PyCall; pyimport_conda("scipy", "scipy")' # retry
-    - |
-      julia --project -e '
-        using Pkg;
-        Pkg.instantiate();
-        Pkg.build();
-        Pkg.test(; coverage=true);'
-  tags:
-    - nvidia
-    - cuda_10.2
-
-test:dev:
-  extends:
-    - .julia:nightly
-    - .test
-  allow_failure: true
-  variables:
-    CI_VERSION_TAG: 'v1.6'
-  script:
-    - |
-      julia --project -e '
-        using Pkg; Pkg.add("PyCall");
-        Pkg.instantiate();
-        ENV["PYTHON"] = ""; Pkg.build();
-        using PyCall; pyimport_conda("scipy", "scipy")'
-    - julia --project -e 'using PyCall; pyimport_conda("scipy", "scipy")' # retry
-    - |
-      julia --project -e '
-        using Pkg;
-        Pkg.instantiate();
-        Pkg.build();
-        Pkg.test(; coverage=true);'
-  tags:
-    - nvidia
-    - cuda_10.2
-
-coverage:
-  stage: post
-  extends:
-    - .julia:1.5
-  script:
-    - julia -e 'using Pkg;
-                Pkg.instantiate();
-                Pkg.add("Coverage")'
-    - julia -e 'using Coverage;
-                cl, tl = get_summary(process_folder());
-                println("(", cl/tl*100, "%) covered")'
-  coverage: '/\(\d+.\d+\%\) covered/'
-
-# pages:
-#   stage: deploy
-#   script:
-#     - mv docs/build public
-#   artifacts:
-#     paths:
-#     - public
-#   only:
-#     - master
