### 2021-04-01 7d2820a
Add tests with Julia 1.6 in github actions
