+++ .github/workflows/ci.yml
@@ -10,8 +10,8 @@ jobs:
       fail-fast: false
       matrix:
         version:
-          - '1.4'
           - '1.5'
+          - '1.6'
           - 'nightly'
         os:
           - ubuntu-latest
