### 2021-04-02 66e33c3
Add ScatterNNlib namespace for gather method
