+++ src/pool.jl
@@ -69,14 +69,14 @@ end
 Dims(xs::AbstractArray{Int}, us::AbstractArray) = Dims(size(xs), size(us))
 
 @adjoint sumpool(cluster::AbstractArray{Int}, X::AbstractArray{T}) where {T<:Real} =
-    sumpool(cluster, X), Δ -> (nothing, gather(zero(Δ)+Δ, cluster))
+    sumpool(cluster, X), Δ -> (nothing, ScatterNNlib.gather(zero(Δ)+Δ, cluster))
 @adjoint subpool(cluster::AbstractArray{Int}, X::AbstractArray{T}) where {T<:Real} =
-    subpool(cluster, X), Δ -> (nothing, -gather(zero(Δ)+Δ, cluster))
+    subpool(cluster, X), Δ -> (nothing, -ScatterNNlib.gather(zero(Δ)+Δ, cluster))
 
 @adjoint function prodpool(cluster::Array{Int}, X::Array{T}) where {T<:Real}
     prodpool(cluster, X), function (Δ)
         rev_cluster = ScatterNNlib.gather_indices(cluster)
-        ∇X = gather(zero(Δ)+Δ, cluster)
+        ∇X = ScatterNNlib.gather(zero(Δ)+Δ, cluster)
         @inbounds for ind = CartesianIndices(cluster)
             inds = filter(x -> x != ind, rev_cluster[cluster[ind]])
             for i = 1:size(X, 1)
@@ -90,7 +90,7 @@ end
 @adjoint function divpool(cluster::Array{Int}, X::Array{T}) where {T<:Real}
     divpool(cluster, X), function (Δ)
         rev_cluster = ScatterNNlib.gather_indices(cluster)
-        ∇X = -gather(zero(Δ)+Δ, cluster) ./ X.^2
+        ∇X = -ScatterNNlib.gather(zero(Δ)+Δ, cluster) ./ X.^2
         @inbounds for ind = CartesianIndices(cluster)
             inds = filter(x -> x != ind, rev_cluster[cluster[ind]])
             for i = 1:size(X, 1)
@@ -104,7 +104,7 @@ end
 @adjoint function maxpool(cluster::Array{Int}, X::Array{T}) where {T<:Real}
     max = maxpool(cluster, X)
     max, function (Δ)
-       Δu = (X .== gather(max, cluster)) .* gather(zero(Δ)+Δ, cluster)
+       Δu = (X .== ScatterNNlib.gather(max, cluster)) .* ScatterNNlib.gather(zero(Δ)+Δ, cluster)
        (nothing, Δu)
     end
 end
@@ -112,7 +112,7 @@ end
 @adjoint function minpool(cluster::Array{Int}, X::Array{T}) where {T<:Real}
     min = minpool(cluster, X)
     min, function (Δ)
-       Δu = (X .== gather(min, cluster)) .* gather(zero(Δ)+Δ, cluster)
+       Δu = (X .== ScatterNNlib.gather(min, cluster)) .* ScatterNNlib.gather(zero(Δ)+Δ, cluster)
        (nothing, Δu)
     end
 end
@@ -120,7 +120,7 @@ end
 @adjoint function meanpool(cluster::AbstractArray{Int}, X::AbstractArray{T}) where {T<:Real}
     m = meanpool(cluster, X)
     m, function (Δ)
-        ΔX = gather(zero(Δ)+Δ, cluster)
+        ΔX = ScatterNNlib.gather(zero(Δ)+Δ, cluster)
         counts = zero.(cluster)
         @inbounds for i = 1:size(m, 2)
             counts += sum(cluster.==i) * (cluster.==i)
