+++ src/cuda/pool.jl
@@ -80,7 +80,7 @@ meanpool(cluster::Array{Int}, X::CuArray{T}, c::Integer=length(Set(cluster))) wh
 @adjoint function prodpool(cluster::CuArray{Int}, X::CuArray{T}) where {T<:Real}
     prodpool(cluster, X), function (Δ)
         rev_cluster = ScatterNNlib.gather_indices(cluster)
-        ∇X = gather(zero(Δ)+Δ, cluster)
+        ∇X = ScatterNNlib.gather(zero(Δ)+Δ, cluster)
         @inbounds for ind = CartesianIndices(cluster)
             ind = Tuple(ind)
             inds = filter(x -> x != ind, rev_cluster[cluster[ind...]])
@@ -99,7 +99,7 @@ end
 @adjoint function divpool(cluster::CuArray{Int}, X::CuArray{T}) where {T<:Real}
     divpool(cluster, X), function (Δ)
         rev_cluster = ScatterNNlib.gather_indices(cluster)
-        ∇X = -gather(zero(Δ)+Δ, cluster)
+        ∇X = -ScatterNNlib.gather(zero(Δ)+Δ, cluster)
         ∇X ./= X.^2
         @inbounds for ind = CartesianIndices(cluster)
             ind = Tuple(ind)
@@ -119,8 +119,8 @@ end
 @adjoint function maxpool(cluster::CuArray{Int}, X::CuArray{T}) where {T<:Real}
     max = maxpool(cluster, X)
     max, function (Δ)
-       Δu = (X .== gather(max, cluster))
-       Δu .*= gather(zero(Δ)+Δ, cluster)
+       Δu = (X .== ScatterNNlib.gather(max, cluster))
+       Δu .*= ScatterNNlib.gather(zero(Δ)+Δ, cluster)
        (nothing, Δu)
     end
 end
@@ -128,8 +128,8 @@ end
 @adjoint function minpool(cluster::CuArray{Int}, X::CuArray{T}) where {T<:Real}
     min = minpool(cluster, X)
     min, function (Δ)
-       Δu = (X .== gather(min, cluster))
-       Δu .*= gather(zero(Δ)+Δ, cluster)
+       Δu = (X .== ScatterNNlib.gather(min, cluster))
+       Δu .*= ScatterNNlib.gather(zero(Δ)+Δ, cluster)
        (nothing, Δu)
     end
 end
