### 2021-04-02 0fe196d
Add ScatterNNlib namespace for scatter methods
