+++ src/cuda/pool.jl
@@ -2,7 +2,7 @@ function sumpool(cluster::CuArray{Int}, X::CuArray{T},
                  c::Integer=length(Set(cluster))) where {T<:Real}
     dims = Dims(cluster, X)
     Y = CUDA.zeros(T, dims.us_dims[1], c)
-    scatter_add!(Y, X, cluster)
+    ScatterNNlib.scatter_add!(Y, X, cluster)
     Y
 end
 
@@ -13,7 +13,7 @@ function subpool(cluster::CuArray{Int}, X::CuArray{T},
                  c::Integer=length(Set(cluster))) where {T<:Real}
     dims = Dims(cluster, X)
     Y = CUDA.zeros(T, dims.us_dims[1], c)
-    scatter_sub!(Y, X, cluster)
+    ScatterNNlib.scatter_sub!(Y, X, cluster)
     Y
 end
 
@@ -24,7 +24,7 @@ function prodpool(cluster::CuArray{Int}, X::CuArray{T},
                   c::Integer=length(Set(cluster))) where {T<:Real}
     dims = Dims(cluster, X)
     Y = CUDA.ones(T, dims.us_dims[1], c)
-    scatter_mul!(Y, X, cluster)
+    ScatterNNlib.scatter_mul!(Y, X, cluster)
     Y
 end
 
@@ -36,7 +36,7 @@ function divpool(cluster::CuArray{Int}, X::CuArray{T},
     dims = Dims(cluster, X)
     FT = float(T)
     Y = CUDA.ones(FT, dims.us_dims[1], c)
-    scatter_div!(Y, FT.(X), cluster)
+    ScatterNNlib.scatter_div!(Y, FT.(X), cluster)
     Y
 end
 
@@ -47,7 +47,7 @@ function maxpool(cluster::CuArray{Int}, X::CuArray{T},
                  c::Integer=length(Set(cluster))) where {T<:Real}
     dims = Dims(cluster, X)
     Y = CUDA.fill(typemin(T), dims.us_dims[1], c)
-    scatter_max!(Y, X, cluster)
+    ScatterNNlib.scatter_max!(Y, X, cluster)
     Y
 end
 
@@ -58,7 +58,7 @@ function minpool(cluster::CuArray{Int}, X::CuArray{T},
                  c::Integer=length(Set(cluster))) where {T<:Real}
     dims = Dims(cluster, X)
     Y = CUDA.fill(typemax(T), dims.us_dims[1], c)
-    scatter_min!(Y, X, cluster)
+    ScatterNNlib.scatter_min!(Y, X, cluster)
     Y
 end
 
@@ -70,7 +70,7 @@ function meanpool(cluster::CuArray{Int}, X::CuArray{T},
     dims = Dims(cluster, X)
     FT = float(T)
     Y = CUDA.zeros(FT, dims.us_dims[1], c)
-    scatter_mean!(Y, FT.(X), cluster)
+    ScatterNNlib.scatter_mean!(Y, FT.(X), cluster)
     Y
 end
 
