### 2021-04-02 2a88f51
fix type feilds name changed casued by updating to Flux v0.12
