### 2021-04-02 6289e6f
fix cuda test as well
