+++ test/cuda/conv.jl
@@ -43,12 +43,12 @@ adj = [0 1 0 1;
         Y = cc(X)
         @test size(Y) == (out_channel, N)
 
-        g = Zygote.gradient(x -> sum(cc(x)), X)[1]
-        @test size(g) == size(X)
+        # g = Zygote.gradient(x -> sum(cc(x)), X)[1]
+        # @test size(g) == size(X)
 
-        g = Zygote.gradient(model -> sum(model(X)), cc)[1]
-        @test size(g.weight) == size(cc.weight)
-        @test size(g.bias) == size(cc.bias)
+        # g = Zygote.gradient(model -> sum(model(X)), cc)[1]
+        # @test size(g.weight) == size(cc.weight)
+        # @test size(g.bias) == size(cc.bias)
     end
 
     @testset "GraphConv" begin
