+++ README.md
@@ -1,15 +1,13 @@
 # GeometricFlux.jl
 
 <p align="center">
-<img width="400px" src="https://github.com/yuehhua/GeometricFlux.jl/raw/master/logos/logo.png"/>
+<img width="400px" src="https://github.com/FluxML/GeometricFlux.jl/raw/master/logos/logo.png"/>
 </p>
 
-[![](https://img.shields.io/badge/docs-stable-blue.svg)](https://yuehhua.github.io/GeometricFlux.jl/stable)
-[![](https://img.shields.io/badge/docs-dev-blue.svg)](https://yuehhua.github.io/GeometricFlux.jl/dev)
-[![Build Status](https://travis-ci.org/yuehhua/GeometricFlux.jl.svg?branch=master)](https://travis-ci.org/yuehhua/GeometricFlux.jl)
-[![codecov](https://codecov.io/gh/yuehhua/GeometricFlux.jl/branch/master/graph/badge.svg)](https://codecov.io/gh/yuehhua/GeometricFlux.jl)
-[![pipeline status](https://gitlab.com/JuliaGPU/GeometricFlux-jl/badges/master/pipeline.svg)](https://gitlab.com/JuliaGPU/GeometricFlux-jl/commits/master)
-[![coverage report](https://gitlab.com/JuliaGPU/GeometricFlux-jl/badges/master/coverage.svg)](https://gitlab.com/JuliaGPU/GeometricFlux-jl/commits/master)
+[![](https://img.shields.io/badge/docs-stable-blue.svg)](https://fluxml.ai/GeometricFlux.jl/stable)
+[![](https://img.shields.io/badge/docs-dev-blue.svg)](https://fluxml.ai/GeometricFlux.jl/dev)
+![](https://github.com/FluxML/GeometricFlux.jl/actions/workflows/ci.yml/badge.svg)
+[![codecov](https://codecov.io/gh/FluxML/GeometricFlux.jl/branch/master/graph/badge.svg)](https://codecov.io/gh/FluxML/GeometricFlux.jl)
 
 GeometricFlux is a geometric deep learning library for [Flux](https://github.com/FluxML/Flux.jl). This library aims to be compatible with packages from [JuliaGraphs](https://github.com/JuliaGraphs) ecosystem and have support of CUDA GPU acceleration with [CUDA](https://github.com/JuliaGPU/CUDA.jl). Message passing scheme is implemented as a flexbile framework and fused with Graph Network block scheme. GeometricFlux is compatible with other packages that are composable with Flux.
 
