### 2021-04-04 c7bb7ab
CompatHelper: bump compat for "GraphSignals" to "0.2"
