+++ Project.toml
@@ -29,7 +29,7 @@ DataStructures = "0.17 - 0.18"
 FillArrays = "0.8 - 0.11"
 Flux = "0.10 - 0.12"
 GraphMLDatasets = "0.1"
-GraphSignals = "0.1"
+GraphSignals = "0.1, 0.2"
 LightGraphs = "1.3"
 MetaGraphs = "0.6"
 Reexport = "0.2, 1.0"
