### 2021-04-09 #70 @DhairyaLGandhi
Thanks @yuehhua

Here are some thoughts on the paper that would be good to be addressed

1. Comparison of capabilities with other similar software - there needs to be a more complete feature comparison taking multiple such packages into account
2. The references don't show up correctly
3. The motivation behind the ScatterNNlib package and what it borrows from the graphs ecosystem isnt established
4. There should be some examples of the modeling achieved by GeometricFlux. While one can conceivably assume some kinds of popular graph learning possibilities, it would be important to state the kinds of models that it has already been used to build (with examples) and ones there is support for
5. How is the Flux and CUDA integration handled? Section 4 needs some clarity.
6. What is the future direction for this package and how does its current flexibility compare with other packages beyond pytorch-scatter
