### 2021-04-10 a20bed2
CompatHelper: bump compat for "CUDA" to "3.0"
