+++ Project.toml
@@ -24,7 +24,7 @@ Zygote = "e88e6eb3-aa80-5325-afca-941959d7151f"
 ZygoteRules = "700de1a5-db45-46bc-99cf-38207098b444"
 
 [compat]
-CUDA = "1.2 - 2.6"
+CUDA = "1.2 - 2.6, 3.0"
 DataStructures = "0.17 - 0.18"
 FillArrays = "0.8 - 0.11"
 Flux = "0.10 - 0.12"
