### 2021-05-12 5af0020
replace comprehension with map
