+++ test/utils.jl
@@ -10,11 +10,6 @@ ef = rand(5, E)
 gf = rand(7)
 
 @testset "utils" begin
-    @testset "topk_index" begin
-        X = [8,7,6,5,4,3,2,1]
-        @test topk_index(X, 4) == [1,2,3,4]
-        @test topk_index(X', 4) == [1,2,3,4]
-    end
 
     @testset "bypass_graph" begin
         fg = FeaturedGraph(adj, nf=nf, ef=ef, gf=gf)
