+++ test/layers/pool.jl
@@ -28,4 +28,10 @@ X = Array(reshape(1:24, 2, 3, 4))
             @test size(y) == (in_channel, k)
         end
     end
+    
+    @testset "topk_index" begin
+        X = [8,7,6,5,4,3,2,1]
+        @test topk_index(X, 4) == [1,2,3,4]
+        @test topk_index(X', 4) == [1,2,3,4]
+    end
 end
