### 2021-05-12 9669d89
refactor pool function into scatter and accept operator directly
