+++ test/cuda/pool.jl
@@ -4,140 +4,112 @@ X = CuArray(reshape(1:24, 2, 3, 4))
 @testset "cuda/pool" begin
     for T = [UInt32, UInt64]
         @testset "$(T)" begin
-            @testset "sumpool" begin
+            @testset "+" begin
                 y = [40, 44, 12, 14, 36, 38, 16, 18, 40, 42]
                 y = reshape(y, 2, 5)
-                @test sumpool(CuArray{Int64}(cluster), T.(X)) == T.(y)
-                @test pool(:add, CuArray{Int64}(cluster), T.(X)) == T.(y)
-                @test sumpool(cluster, T.(X)) == T.(y)
-                @test pool(:add, cluster, T.(X)) == T.(y)
+                @test GeometricFlux.scatter(+, CuArray{Int64}(cluster), T.(X)) == T.(y)
+                @test GeometricFlux.scatter(+, cluster, T.(X)) == T.(y)
             end
 
-            @testset "maxpool" begin
+            @testset "max" begin
                 y = [19, 20, 9, 10, 21, 22, 11, 12, 23, 24]
                 y = reshape(y, 2, 5)
-                @test maxpool(CuArray{Int64}(cluster), T.(X)) == T.(y)
-                @test pool(:max, CuArray{Int64}(cluster), T.(X)) == T.(y)
-                @test maxpool(cluster, T.(X)) == T.(y)
-                @test pool(:max, cluster, T.(X)) == T.(y)
+                @test GeometricFlux.scatter(max, CuArray{Int64}(cluster), T.(X)) == T.(y)
+                @test GeometricFlux.scatter(max, cluster, T.(X)) == T.(y)
             end
 
-            @testset "minpool" begin
+            @testset "min" begin
                 y = [1, 2, 3, 4, 15, 16, 5, 6, 17, 18]
                 y = reshape(y, 2, 5)
-                @test minpool(CuArray{Int64}(cluster), T.(X)) == T.(y)
-                @test pool(:min, CuArray{Int64}(cluster), T.(X)) == T.(y)
-                @test minpool(cluster, T.(X)) == T.(y)
-                @test pool(:min, cluster, T.(X)) == T.(y)
+                @test GeometricFlux.scatter(min, CuArray{Int64}(cluster), T.(X)) == T.(y)
+                @test GeometricFlux.scatter(min, cluster, T.(X)) == T.(y)
             end
         end
     end
 
     for T = [Int32, Int64]
         @testset "$(T)" begin
-            @testset "sumpool" begin
+            @testset "+" begin
                 y = [40, 44, 12, 14, 36, 38, 16, 18, 40, 42]
                 y = reshape(y, 2, 5)
-                @test sumpool(CuArray{Int64}(cluster), T.(X)) == T.(y)
-                @test pool(:add, CuArray{Int64}(cluster), T.(X)) == T.(y)
-                @test sumpool(cluster, T.(X)) == T.(y)
-                @test pool(:add, cluster, T.(X)) == T.(y)
+                @test GeometricFlux.scatter(+, CuArray{Int64}(cluster), T.(X)) == T.(y)
+                @test GeometricFlux.scatter(+, cluster, T.(X)) == T.(y)
             end
 
-            @testset "subpool" begin
+            @testset "-" begin
                 y = [-40, -44, -12, -14, -36, -38, -16, -18, -40, -42]
                 y = reshape(y, 2, 5)
-                @test subpool(CuArray{Int64}(cluster), T.(X)) == T.(y)
-                @test pool(:sub, CuArray{Int64}(cluster), T.(X)) == T.(y)
-                @test subpool(cluster, T.(X)) == T.(y)
[…]
