+++ src/layers/conv.jl
@@ -1,6 +1,3 @@
-const AGGR2STR = Dict{Symbol,String}(:add => "∑", :sub => "-∑", :mul => "∏", :div => "1/∏",
-                                     :max => "max", :min => "min", :mean => "𝔼[]")
-
 """
     GCNConv([graph, ]in=>out)
     GCNConv([graph, ]in=>out, σ)
@@ -186,7 +183,7 @@ the layer instead of only the features.
 - `out`: the dimension of output features.
 - `bias::Bool=true`: keyword argument, whether to learn the additive bias.
 - `σ=identity`: activation function.
-- `aggr::Symbol=:add`: an aggregate function applied to the result of message function. `:add`, `:max` and `:mean` are available.
+- `aggr=+`: an aggregate function applied to the result of message function. `+`, `max` and `mean` are available.
 """
 struct GraphConv{V<:AbstractFeaturedGraph,T} <: MessagePassing
     fg::V
@@ -194,11 +191,11 @@ struct GraphConv{V<:AbstractFeaturedGraph,T} <: MessagePassing
     weight2::AbstractMatrix{T}
     bias::AbstractVector{T}
     σ
-    aggr::Symbol
+    aggr
 end
 
 function GraphConv(el::AbstractVector{<:AbstractVector{<:Integer}},
-                   ch::Pair{<:Integer,<:Integer}, σ=identity, aggr=:add;
+                   ch::Pair{<:Integer,<:Integer}, σ=identity, aggr=+;
                    init = glorot_uniform, bias::Bool=true, T::DataType=Float32)
     w1 = T.(init(ch[2], ch[1]))
     w2 = T.(init(ch[2], ch[1]))
@@ -207,7 +204,7 @@ function GraphConv(el::AbstractVector{<:AbstractVector{<:Integer}},
     GraphConv(fg, w1, w2, b, σ, aggr)
 end
 
-function GraphConv(adj::AbstractMatrix, ch::Pair{<:Integer,<:Integer}, σ=identity, aggr=:add;
+function GraphConv(adj::AbstractMatrix, ch::Pair{<:Integer,<:Integer}, σ=identity, aggr=+;
                    init = glorot_uniform, bias::Bool=true, T::DataType=Float32)
     w1 = T.(init(ch[2], ch[1]))
     w2 = T.(init(ch[2], ch[1]))
@@ -216,7 +213,7 @@ function GraphConv(adj::AbstractMatrix, ch::Pair{<:Integer,<:Integer}, σ=identi
     GraphConv(fg, w1, w2, b, σ, aggr)
 end
 
-function GraphConv(ch::Pair{<:Integer,<:Integer}, σ=identity, aggr=:add;
+function GraphConv(ch::Pair{<:Integer,<:Integer}, σ=identity, aggr=+;
                    init = glorot_uniform, bias::Bool=true, T::DataType=Float32)
     w1 = T.(init(ch[2], ch[1]))
     w2 = T.(init(ch[2], ch[1]))
@@ -234,14 +231,14 @@ function (gc::GraphConv)(X::AbstractMatrix)
     Zygote.ignore() do
         GraphSignals.check_num_node(g, X)
     end
-    _, X = propagate(gc, adjacency_list(g), Fill(0.f0, 0, ne(g)), X, :add)
+    _, X = propagate(gc, adjacency_list(g), Fill(0.f0, 0, ne(g)), X, +)
     X
 end
 function (g::GraphConv)(fg::FeaturedGraph)
     Zygote.ignore() do
         GraphSignals.check_num_node(graph(fg), node_feature(fg))
     end
-    propagate(g, fg, :add)
+    propagate(g, fg, +)
 end
 
[…]
