### 2021-05-12 93aad43
fix scatter for cuda
