### 2021-05-16 107f75f
add doc for bypass_graph
