+++ dev/search_index.js
@@ -1,3 +1,3 @@
 var documenterSearchIndex = {"docs":
-[{"location":"basics/passgraph/#Graph-passing-1","page":"Graph passing","title":"Graph passing","text":"","category":"section"},{"location":"basics/passgraph/#","page":"Graph passing","title":"Graph passing","text":"Graph is an input data structure for graph neural network. Passing graph into GNN layer can have different behaviors. If graph remains fixed across samples, that is, all samples utilize the same graph structure, a static graph is used. Graphs can be carried within FeaturedGraph to provide variable graph to GNN layer. Users have the flexibility to pick a adequate approach for their own needs.","category":"page"},{"location":"basics/passgraph/#Static-graph-1","page":"Graph passing","title":"Static graph","text":"","category":"section"},{"location":"basics/passgraph/#","page":"Graph passing","title":"Graph passing","text":"A static graph is used to reduce redundant computation during passing through layers. A static graph can be set in graph convolutional layers in prior such that graph in layers is used first for computation. A adjacency matrix adj_mat is given to represent a graph and is put into a graph convolutional layer as follow:","category":"page"},{"location":"basics/passgraph/#","page":"Graph passing","title":"Graph passing","text":"GCNConv(adj_mat, feat=>h1, relu)","category":"page"},{"location":"basics/passgraph/#","page":"Graph passing","title":"Graph passing","text":"Simple(Di)Graph, SimpleWeighted(Di)Graph or Meta(Di)Graph provided by LightGraphs, SimpleWeightedGraphs and MetaGraphs, respectively, are acceptable for passing to layer as a static graph. A adjacency list is also accepted in the type of Vector{Vector}.","category":"page"},{"location":"basics/passgraph/#Variable-graph-1","page":"Graph passing","title":"Variable graph","text":"","category":"section"},{"location":"basics/passgraph/#","page":"Graph passing","title":"Graph passing","text":"A variable graph is supported by FeaturedGraph. Each FeaturedGraph contains different graph structure and its features. Data of FeaturedGraph are feed directly to graph convolutional layer or graph neural network to let each feature be learn on different graph structure. A adjacency matrix adj_mat is given to construct a FeaturedGraph as follow:","category":"page"},{"location":"basics/passgraph/#","page":"Graph passing","title":"Graph passing","text":"FeaturedGraph(adj_mat, features)","category":"page"},{"location":"basics/passgraph/#","page":"Graph passing","title":"Graph passing","text":"Simple(Di)Graph, SimpleWeighted(Di)Graph or Meta(Di)Graph provided by LightGraphs, SimpleWeightedGraphs and MetaGraphs, respectively, are acceptable for constructing a FeaturedGraph. A adjacency list is also accepted, too.","category":"page"},{"location":"basics/passgraph/#Cached-graph-in-layers-1","page":"Graph passing","title":"Cached graph in layers","text":"","category":"section"},{"location":"basics/passgraph/#","page":"Graph passing","title":"Graph passing","text":"While a variable graph is given by FeaturedGraph, a GNN layer don't need a static graph anymore. Besides taking off the static graph from arguments of a layer, remember to turn off the cache mechanism. A cache mechanism is designed to cache static graph to reduce computation. A cached graph is gotten from layer and computation is then performed. For each time, it will assign current computed graph back to layer. Assignment operation is not differentiable, so we must turn off the cache mechanism as follow:","category":"page"},{"location":"basics/passgraph/#","page":"Graph passing","title":"Graph passing","text":"GCNConv(feat=>h1, relu, cach
[…]
