### 2021-05-22 #70 @jpfairbanks
1. Writing is telegraphic.
2. Introduction does not motivate graph learning or properly contextualize it in the tradition of graph algorithms or graph data mining
3. Benchmarks refer only to the message passing primitives and not to any learning algorithm or accuracy on a dataset.
4. I don't see anything geometric in this paper, there is no implementation of spaces or embedded graphs or anything else geometric. Graphs are combinatorial objects, not geometric structures. Graphs can be viewed as finite metric spaces, but this work does not engage with that viewpoint. The shortest path metric is not directly being used in the computation of the message passing microbenchmark.
5. There is no learning in this paper, Section 4 states that the GNN layers are compatible with flux layers. I would expect this to be demonstrated with an end-to-end example using Flux to solve a classification or regression problem.

Overall, I think this is a paper reporting a micro-benchmark of message passing on graphs. I would advise a revise and resubmit.
