### 2021-05-23 #70 @ranjanan
You’re welcome to just update the latex source with changes or a new paper
from the ground up too. We’ll just use whedon to regenerate your paper
whenever you’re done.

On Sat, 22 May 2021 at 8:24 PM, Yueh-Hua Tu ***@***.***>
wrote:

> If I'd like to revise this work, what should I do next?
>
