+++ src/layers/conv.jl
@@ -1,16 +1,16 @@
 """
-    GCNConv([graph, ]in=>out)
-    GCNConv([graph, ]in=>out, σ)
+    GCNConv([graph, ]in=>out[, σ=identity]; bias=true)
 
 Graph convolutional layer.
 
 # Arguments
-- `graph`: should be a adjacency matrix, `SimpleGraph`, `SimpleDiGraph` (from LightGraphs)
-or `SimpleWeightedGraph`, `SimpleWeightedDiGraph` (from SimpleWeightedGraphs). Is optionnal so you can give a `FeaturedGraph` to
+
+- `graph`: Should be a adjacency matrix, `SimpleGraph`, `SimpleDiGraph` (from LightGraphs) or `SimpleWeightedGraph`, `SimpleWeightedDiGraph` (from SimpleWeightedGraphs). Is optionnal so you can give a `FeaturedGraph` to
 the layer instead of only the features.
-- `in`: the dimension of input features.
-- `out`: the dimension of output features.
-- `bias::Bool=true`: keyword argument, whether to learn the additive bias.
+- `in`: The dimension of input features.
+- `out`: The dimension of output features.
+- `σ`: Activation function.
+- `bias::Bool`: Keyword argument, whether to learn the additive bias.
 
 Data should be stored in (# features, # nodes) order.
 For example, a 1000-node graph each node of which poses 100 features is constructed.
@@ -23,7 +23,7 @@ struct GCNConv{T,F,S<:AbstractFeaturedGraph}
     fg::S
 end
 
-function GCNConv(fg::AbstractFeaturedGraph, ch::Pair{<:Integer,<:Integer}, σ = identity;
+function GCNConv(fg::AbstractFeaturedGraph, ch::Pair{<:Integer,<:Integer}, σ=identity;
                  init=glorot_uniform, T::DataType=Float32, bias::Bool=true)
     b = bias ? T.(init(ch[2])) : zeros(T, ch[2])
     GCNConv(T.(init(ch[2], ch[1])), b, σ, fg)
@@ -75,18 +75,18 @@ end
 
 
 """
-    ChebConv([graph, ]in=>out, k)
+    ChebConv([graph, ]in=>out, k; bias=true)
 
 Chebyshev spectral graph convolutional layer.
 
 # Arguments
-- `graph`: should be a adjacency matrix, `SimpleGraph`, `SimpleDiGraph` (from LightGraphs) or `SimpleWeightedGraph`,
-`SimpleWeightedDiGraph` (from SimpleWeightedGraphs). Is optionnal so you can give a `FeaturedGraph` to
+
+- `graph`: Should be a adjacency matrix, `SimpleGraph`, `SimpleDiGraph` (from LightGraphs) or `SimpleWeightedGraph`, `SimpleWeightedDiGraph` (from SimpleWeightedGraphs). Is optionnal so you can give a `FeaturedGraph` to
 the layer instead of only the features.
-- `in`: the dimension of input features.
-- `out`: the dimension of output features.
-- `k`: the order of Chebyshev polynomial.
-- `bias::Bool=true`: keyword argument, whether to learn the additive bias.
+- `in`: The dimension of input features.
+- `out`: The dimension of output features.
+- `k`: The order of Chebyshev polynomial.
+- `bias::Bool`: Keyword argument, whether to learn the additive bias.
 """
 struct ChebConv{T,S<:AbstractFeaturedGraph}
     weight::AbstractArray{T,3}
@@ -98,7 +98,7 @@ struct ChebConv{T,S<:AbstractFeaturedGraph}
 end
 
 function ChebConv(fg::AbstractFeaturedGraph, ch::Pair{<:Integer,<:Integer}, k::Integer;
[…]
