### 2021-05-27 a782aa7
replace get_feature with _view
