+++ src/utils.jl
@@ -1,9 +1,3 @@
-## Get feature with defaults
-
-get_feature(::Nothing, i) = nothing
-get_feature(A::Fill{T,2,Axes}, i::Integer) where {T,Axes} = view(A, :, 1)
-get_feature(A::AbstractMatrix, i::Integer) = view(A, :, i)
-
 """
     bypass_graph(nf_func, ef_func, gf_func)
 
