### 2021-05-27 e544db2
cleanup and move bypass_graph to layers
