+++ src/layers/conv.jl
@@ -1,3 +1,10 @@
+function add_self_loop!(adj::AbstractVector{T}, n::Int=length(adj)) where {T<:AbstractVector}
+    for i = 1:n
+        i in adj[i] || push!(adj[i], i)
+    end
+    adj
+end
+
 """
     GCNConv([graph, ]in=>out)
     GCNConv([graph, ]in=>out, σ)
