+++ test/runtests.jl
@@ -25,13 +25,12 @@ tests = [
     "layers/msgpass",
     "layers/conv",
     "layers/pool",
-    "layers/selector",
+    "layers/misc",
     "models",
     "scatter",
     "graph/simplegraphs",
     "graph/weightedgraphs",
     "graph/metagraphs",
-    "utils",
 ]
 
 if Flux.use_cuda[]
