### 2021-05-27 0199272
refactor constructors for graph conv layers
