+++ src/GeometricFlux.jl
@@ -64,7 +64,7 @@ export
     FeatureSelector,
     bypass_graph,
 
-    # graph/index
+    # utils
     generate_cluster
 
 const IntOrTuple = Union{Integer,Tuple}
@@ -72,8 +72,7 @@ const IntOrTuple = Union{Integer,Tuple}
 include("datasets.jl")
 
 include("scatter.jl")
-
-include("graph/index.jl")
+include("utils.jl")
 
 include("layers/gn.jl")
 include("layers/msgpass.jl")
@@ -83,7 +82,7 @@ include("layers/pool.jl")
 include("models.jl")
 include("layers/misc.jl")
 
-include("graph/simplegraphs.jl")
+include("graphs.jl")
 
 
 using .Datasets
@@ -93,12 +92,6 @@ function __init__()
         include("cuda/scatter.jl")
         include("cuda/msgpass.jl")
         include("cuda/conv.jl")
-    end
-    @require SimpleWeightedGraphs = "47aef6b3-ad0c-573a-a1e2-d07658019622" begin
-        include("graph/weightedgraphs.jl")
-    end
-    @require MetaGraphs = "626554b9-1ddb-594c-aa3c-2596fe9399a5" begin
-        include("graph/metagraphs.jl")
     end
 end
 
