### 2021-05-27 d87215b
merge test cases for graphs
