### 2021-06-29 a83bb51
Implement GINConv layer.
