+++ test/layers/conv.jl
@@ -396,4 +396,32 @@ adj_single_vertex =   T[0. 0. 0. 1.;
             @test size(g.nn.b) == size(ec.nn.b)
         end
     end
+
+    @testset "GINConv" begin
+        X = rand(Float32, in_channel, N)
+        Xt = transpose(rand(Float32, N, in_channel))
+        nn = Flux.Chain(Dense(in_channel, out_channel))
+
+        @testset "layer with graph" begin
+            gc = GINConv(adj)
+            @test size(gc.weight) == (out_channel, in_channel)
+            @test size(gc.bias) == (out_channel,)
+            @test graph(gc.fg) === adj
+
+            Y = gc(X)
+            @test size(Y) == (out_channel, N)
+
+            # Test with transposed features
+            Y = gc(Xt)
+            @test size(Y) == (out_channel, N)
+
+            g = Zygote.gradient(x -> sum(gc(x)), X)[1]
+            @test size(g) == size(X)
+
+            g = Zygote.gradient(model -> sum(model(X)), gc)[1]
+            @test size(g.weight) == size(gc.weight)
+            @test size(g.bias) == size(gc.bias)
+        end
+
+    end
 end
