+++ src/GeometricFlux.jl
@@ -44,6 +44,7 @@ export
     GATConv,
     GatedGraphConv,
     EdgeConv,
+    GINConv,
     message,
     update,
     propagate,
