### 2021-07-02 0f2ae79
migrate scatter to NNlib
