+++ src/scatter.jl
@@ -1,133 +1,0 @@
-function scatter(op::typeof(+), cluster::AbstractArray{Int}, X::AbstractArray{T},
-                 c::Integer=length(Set(cluster))) where {T<:Real}
-    dims = Dims(cluster, X)
-    Y = zeros(T, dims.us_dims[1], c)
-    ScatterNNlib.scatter_add!(Y, X, cluster)
-    Y
-end
-
-function scatter(op::typeof(-), cluster::AbstractArray{Int}, X::AbstractArray{T},
-                 c::Integer=length(Set(cluster))) where {T<:Real}
-    dims = Dims(cluster, X)
-    Y = zeros(T, dims.us_dims[1], c)
-    ScatterNNlib.scatter_sub!(Y, X, cluster)
-    Y
-end
-
-function scatter(op::typeof(*), cluster::AbstractArray{Int}, X::AbstractArray{T},
-                  c::Integer=length(Set(cluster))) where {T<:Real}
-    dims = Dims(cluster, X)
-    Y = ones(T, dims.us_dims[1], c)
-    ScatterNNlib.scatter_mul!(Y, X, cluster)
-    Y
-end
-
-function scatter(op::typeof(/), cluster::AbstractArray{Int}, X::AbstractArray{T},
-                 c::Integer=length(Set(cluster))) where {T<:Real}
-    dims = Dims(cluster, X)
-    FT = float(T)
-    Y = ones(FT, dims.us_dims[1], c)
-    ScatterNNlib.scatter_div!(Y, FT.(X), cluster)
-    Y
-end
-
-function scatter(op::typeof(max), cluster::AbstractArray{Int}, X::AbstractArray{T},
-                 c::Integer=length(Set(cluster))) where {T<:Real}
-    dims = Dims(cluster, X)
-    Y = fill(typemin(T), dims.us_dims[1], c)
-    ScatterNNlib.scatter_max!(Y, X, cluster)
-    Y
-end
-
-function scatter(op::typeof(min), cluster::AbstractArray{Int}, X::AbstractArray{T},
-                 c::Integer=length(Set(cluster))) where {T<:Real}
-    dims = Dims(cluster, X)
-    Y = fill(typemax(T), dims.us_dims[1], c)
-    ScatterNNlib.scatter_min!(Y, X, cluster)
-    Y
-end
-
-function scatter(op::typeof(mean), cluster::AbstractArray{Int}, X::AbstractArray{T},
-                 c::Integer=length(Set(cluster))) where {T<:Real}
-    dims = Dims(cluster, X)
-    FT = float(T)
-    Y = zeros(FT, dims.us_dims[1], c)
-    ScatterNNlib.scatter_mean!(Y, FT.(X), cluster)
-    Y
-end
-
-struct Dims
-    xs_dims::Tuple
-    us_dims::Tuple
-
-    function Dims(xs_dims::Tuple, us_dims::Tuple)
-        @assert xs_dims == us_dims[2:end] "X must have the same latter dimension with cluster."
-        new(xs_dims, us_dims)
-    end
-end
-
-Dims(xs::AbstractArray{Int}, us::AbstractArray) = Dims(size(xs), size(us))
-
-@adjoint scatter(op::typeof(+), cluster::AbstractArray{Int}, X::AbstractArray{T}) where {T<:Real} =
-    scatter(op, cluster, X), Δ -> (nothing, nothing, ScatterNNlib.gather(zero(Δ)+Δ, cluster))
-@adjoint scatter(op::typeof(-), cluster::AbstractArray{Int}, X::AbstractArray{T}) where {T<:Real} =
-    scatter(op, cluster, X), Δ -> (nothing, nothing, -ScatterNNlib.gather(zero(Δ)+Δ, cluster))
-
[…]
