+++ test/runtests.jl
@@ -5,6 +5,7 @@ using Flux: @functor
 using FillArrays
 using GraphSignals
 using LightGraphs: SimpleGraph, SimpleDiGraph, add_edge!, nv, ne
+using NNlib
 using PkgVersion
 using SimpleWeightedGraphs: SimpleWeightedGraph, SimpleWeightedDiGraph, add_edge!
 using SparseArrays: SparseMatrixCSC
@@ -14,7 +15,6 @@ using Zygote
 using Test
 
 cuda_tests = [
-    "cuda/grad",
     "cuda/conv",
     "cuda/msgpass",
 ]
@@ -32,6 +32,7 @@ tests = [
 if Flux.use_cuda[]
     using CUDA
     using Flux: gpu
+    using NNlibCUDA
     append!(tests, cuda_tests)
 else
     @warn "CUDA unavailable, not testing GPU support"
