### 2021-07-04 69534cf
remove redundant check_num_node
