+++ src/layers/gn.jl
@@ -1,67 +1,62 @@
-const AGGR2FUN = Dict((+) => sum, (-) => sum, (*) => prod, (/) => prod,
-                      max => maximum, min => minimum, mean => mean)
-
 _view(::Nothing, i) = nothing
 _view(A::Fill{T,2,Axes}, i) where {T,Axes} = view(A, :, 1)
 _view(A::AbstractMatrix, idx) = view(A, :, idx)
 
+aggregate(aggr::typeof(+), X) = vec(sum(X, dims=2))
+aggregate(aggr::typeof(-), X) = -vec(sum(X, dims=2))
+aggregate(aggr::typeof(*), X) = vec(prod(X, dims=2))
+aggregate(aggr::typeof(/), X) = 1 ./ vec(prod(X, dims=2))
+aggregate(aggr::typeof(max), X) = vec(maximum(X, dims=2))
+aggregate(aggr::typeof(min), X) = vec(minimum(X, dims=2))
+aggregate(aggr::typeof(mean), X) = vec(aggr(X, dims=2))
+
 abstract type GraphNet end
 
-@inline update_edge(gn::T, e, vi, vj, u) where {T<:GraphNet} = e
-@inline update_vertex(gn::T, ē, vi, u) where {T<:GraphNet} = vi
-@inline update_global(gn::T, ē, v̄, u) where {T<:GraphNet} = u
+@inline update_edge(gn::GraphNet, e, vi, vj, u) = e
+@inline update_vertex(gn::GraphNet, ē, vi, u) = vi
+@inline update_global(gn::GraphNet, ē, v̄, u) = u
 
-@inline function update_batch_edge(gn::T, adj, E, V, u) where {T<:GraphNet}
+@inline function update_batch_edge(gn::GraphNet, adj, E, V, u)
     n = size(adj, 1)
     edge_idx = edge_index_table(adj)
     mapreduce(i -> apply_batch_message(gn, i, adj[i], edge_idx, E, V, u), hcat, 1:n)
 end
 
-@inline apply_batch_message(gn::T, i, js, edge_idx, E, V, u) where {T<:GraphNet} =
+@inline apply_batch_message(gn::GraphNet, i, js, edge_idx, E, V, u) =
     mapreduce(j -> update_edge(gn, _view(E, edge_idx[(i,j)]), _view(V, i), _view(V, j), u), hcat, js)
 
-@inline update_batch_vertex(gn::T, Ē, V, u) where {T<:GraphNet} =
+@inline update_batch_vertex(gn::GraphNet, Ē, V, u) =
     mapreduce(i -> update_vertex(gn, _view(Ē, i), _view(V, i), u), hcat, 1:size(V,2))
 
-@inline function aggregate_neighbors(gn::T, aggr, E, accu_edge) where {T<:GraphNet}
+@inline function aggregate_neighbors(gn::GraphNet, aggr, E, accu_edge)
     @assert !iszero(accu_edge) "accumulated edge must not be zero."
     cluster = generate_cluster(E, accu_edge)
     GeometricFlux.scatter(aggr, cluster, E)
 end
 
-@inline function aggregate_neighbors(gn::T, aggr::Nothing, E, accu_edge) where {T<:GraphNet}
+@inline function aggregate_neighbors(gn::GraphNet, aggr::Nothing, E, accu_edge)
     @nospecialize E accu_edge num_V num_E
 end
 
-@inline function aggregate_edges(gn::T, aggr, E) where {T<:GraphNet}
-    u = vec(AGGR2FUN[aggr](E, dims=2))
-    aggr == :sub && (u = -u)
-    aggr == :div && (u = 1 ./ u)
-    u
-end
[…]
