+++ Project.toml
@@ -23,18 +23,18 @@ Zygote = "e88e6eb3-aa80-5325-afca-941959d7151f"
 ZygoteRules = "700de1a5-db45-46bc-99cf-38207098b444"
 
 [compat]
-CUDA = "1.2 - 2.6, 3.0"
-DataStructures = "0.17 - 0.18"
-FillArrays = "0.8 - 0.11"
-Flux = "0.10 - 0.12"
+CUDA = "2.6"
+DataStructures = "0.18"
+FillArrays = "0.8 - 0.12"
+Flux = "0.12"
 GraphMLDatasets = "0.1"
 GraphSignals = "0.1"
 LightGraphs = "1.3"
 MetaGraphs = "0.6"
-Reexport = "0.2, 1.0"
-Requires = "1.0.0"
-ScatterNNlib = "0.1"
-Zygote = "0.4 - 0.6"
+Reexport = "1.1"
+Requires = "1.1"
+ScatterNNlib = "0.1.7"
+Zygote = "0.6"
 ZygoteRules = "0.2"
 julia = "1.4 - 1.6"
 
