### 2021-07-08 88b8b16
refactor test cases
