### 2021-07-08 b5234e1
Remove obsolete `cached` keyword from documentation
