+++ Project.toml
@@ -23,7 +23,7 @@ Zygote = "e88e6eb3-aa80-5325-afca-941959d7151f"
 ZygoteRules = "700de1a5-db45-46bc-99cf-38207098b444"
 
 [compat]
-CUDA = "2.6"
+CUDA = "2.1 - 3.3"
 DataStructures = "0.18"
 FillArrays = "0.8 - 0.12"
 Flux = "0.12"
@@ -36,7 +36,7 @@ Requires = "1.1"
 ScatterNNlib = "0.1.7"
 Zygote = "0.6"
 ZygoteRules = "0.2"
-julia = "1.4 - 1.6"
+julia = "1.5 - 1.6"
 
 [extras]
 PkgVersion = "eebad327-c553-4316-9ea0-9fa01ccd7688"
