+++ .github/workflows/test.yml
@@ -1,7 +1,13 @@
-name: CI
+name: Unit test
+
 on:
-  - push
-  - pull_request
+  create:
+    tags:
+  push:
+    branches:
+      - master
+  pull_request:
+
 jobs:
   test:
     name: Julia ${{ matrix.version }} - ${{ matrix.os }} - ${{ matrix.arch }} - ${{ github.event_name }}
@@ -12,22 +18,27 @@ jobs:
         version:
           - '1.5'
           - '1.6'
-          # - 'nightly'
+          - 'nightly'
         os:
           - ubuntu-latest
           - macOS-latest
           - windows-latest
         arch:
           - x64
+
     steps:
       - uses: actions/checkout@v2
         continue-on-error: ${{ matrix.julia-version == 'nightly' }}
-      - uses: julia-actions/setup-julia@v1
+
+      - name: "Set up Julia"
+        uses: julia-actions/setup-julia@v1
         continue-on-error: ${{ matrix.julia-version == 'nightly' }}
         with:
           version: ${{ matrix.version }}
           arch: ${{ matrix.arch }}
-      - uses: actions/cache@v1
+
+      - name: "Cache artifacts"
+        uses: actions/cache@v1
         continue-on-error: ${{ matrix.julia-version == 'nightly' }}
         env:
           cache-name: cache-artifacts
@@ -38,28 +49,18 @@ jobs:
             ${{ runner.os }}-test-${{ env.cache-name }}-
             ${{ runner.os }}-test-
             ${{ runner.os }}-
-      - uses: julia-actions/julia-buildpkg@v1
+
+      - name: "Build packages"
+        uses: julia-actions/julia-buildpkg@v1
         continue-on-error: ${{ matrix.julia-version == 'nightly' }}
-      - uses: julia-actions/julia-runtest@v1
+
+      - name: "Unit Test"
+        uses: julia-actions/julia-runtest@v1
         continue-on-error: ${{ matrix.julia-version == 'nightly' }}
+
       - uses: julia-actions/julia-processcoverage@v1
-      - uses: codecov/codecov-action@v1
+
+      - name: "Code coverage"
+        uses: codecov/codecov-action@v1
         with:
           file: lcov.info
-  docs:
-    name: Documentation
-    runs-on: ubuntu-latest
-    steps:
-      - uses: actions/checkout@v2
-      - uses: julia-actions/setup-julia@v1
-        with:
-          version: '1'
-      - run: |
-          julia --project=docs -e '
-            using Pkg
-            Pkg.develop(PackageSpec(path=pwd()))
-            Pkg.instantiate()
-            include("docs/make.jl");'
-        env:
-          GITHUB_TOKEN: ${{ secrets.GITHUB_TOKEN }}
-          DOCUMENTER_KEY: ${{ secrets.DOCUMENTER_KEY }}
