### 2021-07-10 a706351
use GraphLaplacians.add_self_loop!
