+++ Project.toml
@@ -8,6 +8,7 @@ CUDA = "052768ef-5323-5732-b1bb-66c8b64840ba"
 DataStructures = "864edb3b-99cc-5e75-8d2d-829cb0a9cfe8"
 FillArrays = "1a297f60-69ca-5386-bcde-b61e274b549b"
 Flux = "587475ba-b771-5e3f-ad9e-33799f191a9c"
+GraphLaplacians = "a1251efa-393a-423f-9d7b-faaecba535dc"
 GraphMLDatasets = "21828b05-d3b3-40ad-870e-a4bc2f52d5e8"
 GraphSignals = "3ebe565e-a4b5-49c6-aed2-300248c3a9c1"
 LightGraphs = "093fc24a-ae57-5d10-9952-331d41423f4d"
@@ -26,6 +27,7 @@ CUDA = "2.1 - 3.3"
 DataStructures = "0.18"
 FillArrays = "0.8 - 0.12"
 Flux = "0.12"
+GraphLaplacians = "0.1"
 GraphMLDatasets = "0.1"
 GraphSignals = "0.2"
 LightGraphs = "1.3"
