+++ Project.toml
@@ -25,7 +25,7 @@ Zygote = "e88e6eb3-aa80-5325-afca-941959d7151f"
 [compat]
 CUDA = "2.1 - 3.3"
 DataStructures = "0.18"
-FillArrays = "0.8 - 0.12"
+FillArrays = "0.11 - 0.12"
 Flux = "0.12"
 GraphLaplacians = "0.1"
 GraphMLDatasets = "0.1"
@@ -37,7 +37,7 @@ NNlibCUDA = "0.1"
 Reexport = "1.1"
 Requires = "1.1"
 Zygote = "0.6"
-julia = "1.5 - 1.6"
+julia = "1.6"
 
 [extras]
 PkgVersion = "eebad327-c553-4316-9ea0-9fa01ccd7688"
